Code/RDGeneral/testDict.cpp

+++ Code/RDGeneral/testDict.cpp
@@ -96,7 +96,7 @@ void testRDAny() {
     bool a = true;
     RDValue v(a);
     TEST_ASSERT(rdvalue_cast<bool>(v) == true);
-    v = (int)10;
+    v = 10;
     TEST_ASSERT(rdvalue_cast<int>(v) == 10);
   }
